Fix up the code.
