+++ tests/functional/adapter/persist_docs/fixtures.py
@@ -1,9 +1,3 @@
-import json
-import os
-import pytest
-
-from dbt.tests.util import run_dbt
-
 _MODELS__MY_FUN_DOCS = """
 {% docs my_fun_doc %}
 name Column description "with double quotes"
@@ -113,66 +107,3 @@ _SEEDS__BASIC = """id,name
 1,Alice
 2,Bob
 """
-
-class TestPersistDocsDelta:
-    @pytest.fixture(scope="class")
-    def models(self):
-        return {
-            "incremental_delta_model.sql": _MODELS__INCREMENTAL_DELTA,
-            "my_fun_docs.md": _MODELS__MY_FUN_DOCS,
-            "table_delta_model.sql": _MODELS__TABLE_DELTA_MODEL,
-            "schema.yml": _PROPERTIES__MODELS
-        }
-
-    @pytest.fixture(scope="class")
-    def seeds(self):
-        return {
-            "seed.csv": _SEEDS__BASIC,
-            "seed.yml": _PROPERTIES__SEEDS
-        }
-
-
-    @pytest.fixture(scope="class")
-    def project_config_update(self):
-        return {
-            'models': {
-                'test': {
-                    '+persist_docs': {
-                        "relation": True,
-                        "columns": True,
-                    },
-                }
-            },
-            'seeds': {
-                'test': {
-                    '+persist_docs': {
-                        "relation": True,
-                        "columns": True,
-                    },
-                    '+file_format': 'delta',
-                    '+quote_columns': True
-                }
-            },
-        }
-
-    def test_delta_comments(self, project):
-        run_dbt(['seed'])
-        run_dbt(['run'])
-
-        for table, whatis in [
-            ('table_delta_model', 'Table'),
-            ('seed', 'Seed'),
-            ('incremental_delta_model', 'Incremental')
-        ]:
-            results = project.run_sql(
-                'describe extended {schema}.{table}'.format(schema=project.test_schema, table=table),
-                fetch='all'
-            )
-
-            for result in results:
-                if result[0] == 'Comment':
-                    assert result[1].startswith(f'{whatis} model description')
-                if result[0] == 'id':
-                    assert result[2].startswith('id Column description')
-                if result[0] == 'name':
-                    assert result[2].startswith('Some stuff here and then a call to')
